extract Login and Healthcheck to their own controllers
in preparation of removing the old angular app
